I did not think that would work.
I've been fiddling with the Delaunay triangulation for ages, and I've finally gotten it to work up to aspect ratios of around 2. Above that, I even managed to get it to work _most_ of the time. I'm not sure how this works at all, or why it doesn't work completely. I'm not sure of anything. Wow this is hard.