Let's make some maps!
I broke the existing code into two parts. Now the "Planet" tab only renders the shape of a planet with insolation, while the "Terrain" tab actually handles the Delaunay triangulation stuff. Now I just need to figure out how to make a map so that I can see the terrain as it appears.