Civi!
I made tectonic plates! The globe is now divided into 12 random shapes. Main issues at the moment: the edges are far too fuzzy (there are too many exclaves), and the tiles near the south pole are obscuring a lot of the inner ones. I have plans to fix both of those. In the near future.