Keep it together
I added a modification to my plate spreading algorithm so that they are far less likely to become non-contiguous. This also makes them much more regularly-shaped, which I think is for the best. It's still hard to see _exactly_ what's going on, because I haven't fixed the thing with the tiles near the singularity yet. That's next. I'm thinking a two-step approach of adding resolution to edges _near_ it, and inverting the tile that _contains_ it.