These are the worst oceans ever.
I implemented plate tectonics in full. Now, all kinds of nonconservative plate boundaries exist, and they each have fun heightmaps that go with them.

The problem is that I'm not getting the kinds of nice symmetric oceans that Earth has in the Atlantic. The faultlines are too jagged, so any continental splits are peppered with floating mountain ranges. I did a bit of local aveaging to try to compensate for this, but it's clear that more needs be done. I'm still trying to figure out exactly what that more will be.