Dayumn that is looooow
I changed it so the user (more or less) directly controls the sea level, rather than the fraction of the planet that is aquatic. This means that the same sea level input always yields relatively similar-looking planets rather than it occasionally changing drastically based on whether the abount of oceanic crust happens to be more than the desired oceanic fraction.