good enough. for now.
I went ahead and implemented the priority thing. It reduced the number of islands in what should be open young ocean by about half. I realize I really don't want to do the directional thing I was thinking of (i.e. having fault effects propogate only in the direction of their relative velocities), so I'm going to leave it like this for now and see how the biomes look.