I made it worse.
There's good news and bad news. The good news is it turns out you can use unicode in JS variable names. Also, I replacced my parametric latitudes with geodetic ones, which I've been meaning to do for a while now, not that that changed anything. The bad news is random seed 1 still has a triangulation error when the day length is set to 1.8, and also my continents are all boring and broken looking, and I suspect that something about my new rift code is to blame. So there's that.