Look out for barbarians
I fleshed out the spreading system so that the rate at which empires expand depends on its personality. I basically wrote a simple little game for them to play. Right now it seems that there's not as much war as I would like, and the empires are too irregularly shaped for my taste, but it'll do for now.

I think that before I can debug much further these blobs will need names. Guess it's time for language construction!