Was this always so slow?
I fixed the shading finally. It's, like, pretty slow. I don't remember it taking more time than the actual terrain generation, especially since it should just be a loop through the tiles. Could be the blurring. In any case, I think it's best to leave it off the political map for now.