Less noticeable, but still prominent
It seems like sites tend to be pretty split on whether they represent "language" with 🌐 or 文/A. I decided to go with 文, because it ensures that it will match the font color even if I change it, and because unlike the Latin "A", it already means "language". Aren't logographies the best? Maybe a little weird since Chinese isn't a language I support, and 文 doesn't mean the same in Japanese bu whatevs.