Maps!!
I implemented the map tab! it's still very basic, obviously, but there are multiple color schemes from which you can choose, and two whole projections! adjusting the mapping system to account for the different topology of the equirectangular projection proved to be far more challenging than I had originally expected, but hopefully now it'll all be a lot more robust.