novi rang
I updated the country colors so that there are more of them, optimized them a little more scientifically, and actually saved the way I generated them. I'll be honest, I feel like they're not that perceptually uniformly distributed; it seems like there's a lot of magentas and blues like I would expect from HSV, but whatever. the math gods have spoken, and I need to not worry about it.