Today, ladies
I implemented a loading spinner! It's only on one of the buttons; I'm going to put it on the rest of them now, but it was actually a decent amount of work. I had to move the generation code out of the callback and into a separate function. I'll probably need to do that anyway when I have it update on tab clicks and on opening.