the most extra loading bar ever
I replaced the spinner with an animation of the dunhamar logo rotating around a globe. I think it's really neat. it is grainier than I would like because GIF only supports two levels of transparency (0 and 100), but I couldn't find any alternative animated format that didn't seem like it would take a ton of work to implement.