we can't have nice things
I figured out ho to make the tabs show properly: turn off the fading. it was neat, but I guess Chrome can't think about terrain generation and fade a tab in at the same time unless I want to figure out web workers (which I don't), and I don't want to wait for the fading to finish before I start the thinking, so best to just abandon the fading entirely so Chrome can concentrate on what's important. it also makes it a bit easier to see the parallels between the different tabs.