if plotly plots in the forest
I figured out a way to get the Plotly sizing to stop being so terrible. it's a bit slower, but I don't think it's a big deal that opening the 3D plot for the first time makes you wait a second. though I should probably make it not do that on subsequent openings.