Ignition
I reversed the order of rules such that assimilation now propogates. I also reversed vowel harmony so the word is determined by the first vowel, not the last. and to do that, I changed voicing assimilation to a normal rule, and added a new syntax to the rules so that they can have proper assimilation like that.