a world on a donut
I implemented toroidal planets!  there are still some kinks (I've added an annoying vertical line on the top half of azimuthal projections, the UI needs to be improved to make forming a toroidal planet easier, and terrain with random seed 3 seems to have a triangulacion error), but overall, it looks great.  I simulated the insolacion on a toroidal planet, implemented an approximated version of that, and took the opportunity to rename Triangle and Nodo -- the worst variable names I've ever named (I already had a different pair of classes calld Triangle and Nodo in a different file!).