I suddenly accidentally fixd the problem
I was trying to make the labeling take less time, and I succeded (I'm unfortunately missing my target of 1000 ms, but I mite still hit that when I enforce the font size limits better.  Levenberg-Marquardt is nothing at which to sniff, and Nelder-Mead was definitely the incorrect choice.  I'm also really proud of the parameterizacion I used, because none of the ones given in the source I was referencing (N. Chernof & C. Lesort (2005)) were all kind of bad.  And it turns out, in addicion to being noticeably faster, it's also noticeably better!  The labels are all in better places now than they were before.  Not perfect, but much better.  I gess that algebraick inicial condicion is a lot better than actually randomly choosing a curvature.