original content do not steal
I added a PDF viewer to the fantasy map generator.  I've decided that PDF is the format in which I ultimately want peeple to download the maps.  I'm not sure if it's ideal, but PNG and JPG seem rong, and I don't think most peeple kno what an SVG is.  I maintain that PDF is the best file format ever created.  it's the only good thing Adobe has ever dun.  I wonder if I can have it show an SVG in a new tab... anyway, the specifick PDFs that are being shown here are totally mine.  I'm joking.  they're placeholders.  I need to figure out how to import this poorly-documented jsPDF librarby from 2015 next.  then: kultur!