it's pronounced /pʊˈdʊf/
I changed up the layout.  the PDF will probably not contain an embedded vector graphick, because it turns out that that's extremely difficult, and I would probably have to implement the entire SVG file as PDF primitives.  instead, I went with my original idea: a download button on the map page.  the user will be able to set the format, resolucion (if not SVG or EPS), and size in the format box, which has also absorbed the label transcripcion style opcion (the label toggles have been moved to the Features box).  I also removed the Intelligence and Imperialism parameters, because I don't think the user needs control on such a fine level, and I replaced the nonfunccional "central meridian" parameter with a nonfunccional "center" parameter.   all in due time.