there's just one thing we need to take care of
I made it center on a country of the user's choice.  there's a problem where it turns out that ids aren't actually unique for some reason, but I'll get to fixing that later.  it works decently for now.

oh, I also found out that the default intellij font has ⟨ɸ⟩ but not ⟨φ⟩, so I switchd to that one (it's the one that's on my keyboard, anyway.

I also defined a new filter funccion to replace the one that Javascript should have but doesn't