I implemented it.
there is now a Bonne projeccion.  it's the default now.  I haven't cropped it to the relevant meridians and parallels yet because it will be hard.  I refactord the eroding segment tree and labelling, as well.  also, the labels suddenly got a lot more many and some of them are extremely small.  idk what's up with that.