swoosh!
I added the beginnings of greebles.  they are not currently random, so they look silly, but that's my next step.  I also refactord some geometry stuff, as per uge.  this does fail sometimes, so I need to make it more robust to self-intersecting curves, since I imagine that is going to start happening from time to time from now on.