this is your half of the earth
I implemented a generic Tree class that is more efficient than TreeMap, so that I could implement the strait skeleton algorithm for polygons, so that I can partition node area among the surrounding edges, so that I can make my greebles not intersect with each other.  they still intersect with themselves sometimes, unfortunadamente, but I haven't yet thaut of an efficient way to prevent that.  for now, it seems to be working fine.

I also went ahead and made it so that it will always list at least 4 countries in the country dropdown, even if the 4 countries it has to show are smaller than it would normally allow.