Selector?  I barely know 'or!
I discoverd a bug where going to the planet tab and then going to the terrain tab and then going to the history tab without changing anything caused it to crash.  I had to make my state tracking more intelligent.  in the process, I decided to go ahead and replace jquery.  my syntax is a little nicer (IMO) and it cashes its queries now!