"a *slite* chance"
I had to remove the entering/exiting system because it was fundamentally flawd.  it could never account for segments that start and end on the edge but that aren't crossings.  I didn't think this would come up, but it turns out it does... specifically in the Mercator projection.  so I replaced it with a more thoro check-each-vertex-then-check-each-midpoint-until-you-get-an-unambiguous-anser system.  I had to get rid of the touchedLoop system as well for similar reasons, but that wasn't really doing much as it turns out.

I also removed some dangerous roundoff stuff.  remember kids, no arithmetic operation is truly reversible.

as a side note, it currently seems to fail when the map is centerd on the country that contains the south pole.