TEST
I set up tests using jest.  for now, I'm only testing util.ts now, but I figure I can always add tests for other files later when I think it's necessary.

I didn't uncover any major problems, but this did prompt me to revise some of the utility functions to be safer, and rewrite some of the documentation that was vague or misleading.

it's not set up to test on push rite now because I'm not convinced it's necessary and I onestly feel bad about using all that computing power.  I know it's probably silly but I think this is fine for now.