niftimal random seeds
I revamped the RNG of the language generator to make it more stable (now the country names don't randomly completely change from one timestep to another) and more robust (now there aren't always multiple countries with the same name).  the big thing here was setting it up so that the generation of a random protoword can take a string as a random seed.