names that seed together bleed together
I made it so the seed increments simultaneusly for the first and last names.  I had previusly added a feature where the last name wouldn't always increment so adjacent names would share a family name but the given names would all be uneke.  but after looking at it for a while I think it just doesn't look good.  so we're back to unique family names as well, and if the user doesn't kno which is which, no one cares.