fix transcription of linguolabial lateral fricative
it was failing to transcribe this sound because it tries to transcribe linguolabial fricatives sometimes by using the bilabial equivalent, and there is no symbol for bilabial lateral fricative.  so I made it treat linguolabial lateral fricatives the same as linguolabial liquids, where it just uses the coronal symbol.  also, i deleted linguolabial trills while I was at it because I don't think it's possible.