fix up slavic-style sibilants
it was weird to me that one of the transcription systems used ⟨tsh⟩ for /t͡ʃ/ but also ⟨c⟩ for /t͡s/.  I'm genuinely not sure what I was thinking there.  I've moved the affricate-c stuff into the slavic-style transcription system (column 2) and switched around some of the retroflex sibilants, and now it feels much more thematicly consistent.