revise .gitignore
the working tree was a bit of a mess; most of the changes in many situations were
minor edits to the HTML files that got multiplied by six because of all the
redundant generated inlized versions.  I think I chose not to ignore them originally
because I was looking toward hosting this on github pages, in which case they would
all need to be in the git repo eventually.  but I was never including the generated
javascript files that are redundant with the typescript files but need to be hosted
eventually.  so I'll figure out the hosting later; for now I just need these gone.

while I was at it, I noticed a similar but much worse issue, which is that I also
wasn't including my dependencies.  on its face that sounds reasonable, but the issue
is that I don't know how to use npm so I downloaded these dependencies from the
internet and didn't clearly document which version I'm using, and I also had to
manually edit a couple of them to make them work with ECMA script 6.  so it's not
at all reasonable for me to ask others (or future me) to recover those dependencies after cloning the repo.  now they're inclued, as they should have been already.