js bad
the flowing wasn't doing correctly so the rivers were *not* tapering monotonicly like they should.  the problem was that not all vertices are in the riverDistance Map anymore, so when it tried to sort all vertices by riverDistance, there were a bunch of undefineds (what's a key error?  I hardly know 'er!) and that caused the list to just not sort correctly at all.  so I excluded endorheic vertices, and now the rivers are much more logical.