make rivers a bit more efficient
I realized I could cut out the step in this process where I sort the rivers by distance from the ocean if I just store the rivers in a stack as I build them out from the ocean.  so I did that.  I don't kno how much time that saved, but it's surely a good thing to have as few n*log(n) processes as possible, rite?  I also made it so vertices with no river have a downstream of null rather than of undefined.