rename Nodo -> Tile, Triangle -> Vertex
I renamed a bunch of classes and variables to reframe the spacial representation of the map as a voronoi plot rather than as a Delaunay triangulation.  this is how it's drawn, after all, and how it's modelled in the vast majority of the code.  it's really only in the very beginning, when it's placing the points on the surface and calculating the adjacency graph, that it treats it at all as a Delaunay triangulation graph.  so it makes the most sense to drop that representation as soon as it's done and switch over to the Voronoi terminology.