just-in-time greebling
it turns out that the thing that was making my code so slow was the fact that it had to greeble every edge between every pair of adjacent tiles before it could even start generating the heightmap.  as it turns out, the vast majority of these edges will never be drawn with greebles, so this was a humungus waste of time (thank goodness, since I would be sad if this slowness was the necessary price of greebles).  by setting it to only greeble an edge right before those greebles get renderd, I've reduced the amount of computation drasticly.  and there's even more room for improvement, since now I will make the greebling scale-dependent!