the same greebling, but now more complicated
I'm setting up to do the scale-dependent greebling, but I don't have it tied to actual scale yet; it's just always going to the same level of detail as before.  I implemented this funny linkedlist-type thing to make it efficient, but I'm realizing now it's way more complicated than it needs to be...
this also involved finding and fixing a bug where I was using Edge.length inconsistently to refer to both the vertex-to-vertex distance and the tile-to-tile distance.