fix nan greebles on Disc
it seems that when I introduced greebles there was a bug that occurred specificly for Disc where it would hit NaN when trying to calculate the intersection with the bounds polygon.  this was specificly because there were zero-length edges in the bounds polygon, but *only* when the coordinate conversions between the edge and tile coordinates were exact.  in other words, it didn't affect toroid and spheroid because the curvature displaced the points slitely so that the edge still had some length.  in fact I probably did that somewhat intentionally not realizing that not all surfaces have curvature.
 real spherist of me.  anyway, I added a test where it actually greebles at finite resolution on a Disc so this exact issue won't come up agen.  as for the actual fix, it now uses the exact endpoint coordinates for those points of the polygon rather than coordinate-converting them at all.

there's another error that's causing Discs to still break, and it looks like it's been there even longer.  *sigh* I'll keep working at it.