less janky paths that go over the edge
maps made on a Disc would often fail because the mapping function occasionally needs to calculate the midpoint between two points, and the way I had implemented lines around the edge of the disc was by setting the latitude of points on the edge to negative infinity.  that causes nans when you try to work with it.  very jank!  not a good idea!  I've revised the outline function now to simply cut the paths at that edge.  this is probably a bit more efficient, but most importantly it doesn't involve using NEGATIVE_INFINITY to represent a special value, which was not intuitive or obvius from how the variables are named.  I had to redefine a "close path" slitely since places where I could once assume the path to be truly closed now sometimes see paths that appear to go off the edge of the map and get clipped.