do one thing and do it well
I refactored the greebling ever so slitely.  I felt it was a bit unclear that Chart.outline was also greebling, and there wasn't really any reason for those to be in the same function.  so now outline just returns lists of Vertexes, and you can choose to greeble and pathify them or not as you please.