scale-dependent rivers
now the flow rate that a river needs to warrant placement on the map is scale-dependent, so as you zoom in more rivers will appear.  I'm not completely satisfied with the algorithm I used to decide which rivers appear; it's all based on flow-rate rite now and the flow-rate threshold goes like the areal scale of the map, which tends to have the result that world maps have no rivers at all, and very local maps have a lot of rivers.  I think ideally I could count up the number of rivers in frame and decide the threshold based on that, but that would be much harder than just scaling the map area.  also this way you can kind of tell how significant a river is based on whether it appears or not, you know?  like a map of bangladesh would show more rivers than a map of mauritania.  anyway, it's fine for now and I think it looks nice.

to make this work I needed to adjust the equirectangular and mercator projections to be true-scale.  previusly I really hadn't given any thaut at all to the scales of the map projections but now they're al true-scale in km until you fit the SVG to whatever sized rectangle.