reasonable voronoi graph on the rim
I made it so that the voronoi graph stops at the rim with edges terminating at the rim on vertices as one would expect, rather than having a few vertices outside of it and then also terminating the edges on null vertices that are actually multiple radii beyond the rim.  to make this work I moved the location calculation out of the constructor into a static function.
in order to effectively test this new vertex placement functionality, I ended up also separating the random tile location selection into a different function from the triangulation and graph-bilding.