rename many files
I've been a little bit annoyed with my filenames for a while now.  so I'm writing a manifest to remind myself what each folder is supposed to be, and taking the opportunity to rename most of them.  in particular I hate abbreviations in contexts like this, so "src" is becoming "source" and "res" is becoming "resources".
we'll see if I like this one better.  I mean, organizing code is just hard.  but I'm feeling good about this scheme.
I do enjoy having the freedom to refactor whatever I want whenever I want.  not that I've ever really workd on a large project with other peeple so maybe it's not so bad.