clean up licensing
I realized that I had both MIT licenses in my code and CC0 licenses on my wobpages, and also that not all of the dependencies were clearly annotated as others' work.  so I cleand all that up.  now everything is under a CC0 license and all of my sources are properly cited.