fix language codes
Japanese is ja, not jp.  that always gets me because Japan is JP, an because of jpdb.io.  anyway, it would have caused problems with screen-readers so I changed it to ja everywhere.  similarly, since pd is not a real language code, I should not be using it in HTML lang attributes.  so I set it up where it now uses "en" in place of "pd" anywhere a real language code is required.