add special tiles to Disc edge
I had some problems trying to implement a new feature relating to the fact that trying to reference the neibor of a Tile adjacent to the rim of the Disc returned null.  I think this is just too widespread a thing for which to use null.  so instead I'm creating a new object that stands in for the void beyond the rim.  working with null is kind of a pain; I wish this were rust so I could use specificly designated dummy values like this more easily.  idk I gess it's not too hard; maybe I should actually just designate dummy values more often.  while I was on the topic I made the result of passing a null World to depict a bit more explicit.