remove TreeMap bloat
this data structure had a really curius feature which was that it kept track of the sum of the values of all its elements, and did so recursively.  it lookd like it was doing something clever and efficient but it was actually going about it in a completely pointlessly complicated way.  this feature really doesn't need to be here and the fact that it is in spite of that annoys me.