rescale arability values
I rescaled all the values so that the arable area of a tile maxes out at the tile area, rather than forest-dwellers being able to farm 300% of their land.  the ratios all changed a little bit because I tried to keep all the numbers round and it's not like they were very fine-tuned to begin with, but I tried to keep it so everything is about the same behavior.