rename dAds to dsdλ
I like this because it reduces the number of times I need to multiply or divide by 2π, and it makes the two surface differential functiosn more similar to each other.  i think I just didn't notice when I first rote this that it was an option to do it this way.  it makes the double partial derivative kind of awkward, tho.