true Equal Earth projection
I've been inconsistently calling this pseudocylindrical map of mine the "equal earth" or "equal area" projection.  obviusly "equal area" is not descriptive, since I also have the Bonne projection.  but it also wasn't really the Equal Earth projection; just like it in that it had a pole line and was equal-area and pseudocylindrical.  well, now I've done some math to make it... not equal to the equal earth projection, but close enuff that you can't distinguish them by eye.  and fully generalized to ellipsoids and tori!  now if only I could figure out what to do when the user selects equal earth on a Disc... I suppose I can always just generalize it to have other standard parallels (tho in that case it becomes teh one true projection on a disc so maybe I shouldn't bother and just say that Disc doesn't get Equal Area)