convert all resource text files to JS
so I started this because I wanted to write tests to validate that the resource files didn't contain any typos, because the current system of I run the code a bunch of times and wait for an error to appear is pretty bad.  but you can't load resources that aren't JS or JSON in tests.  so I decided to convert all those TSV files to JSON.  the problem is you can't import JSON on the client side because of… security? reasons??  or well you're going to be able to soon, but the way to do that that gets around the security rules hasn't been implemented by any browsers (why does this feature even exist if not to be used in browsers???)

so anyway, that's my life story and now all of the resource files are just TypeScript, written such that it will be extremely easy to port them to JSON if that feature ever becomes as real as I feel like it already should have been from the start and if I feel like changing it.  I had trepidations at first because it makes most of them much longer than they were otherwise.  it also I think makes some of them much harder to read for me who knows how all of them are laid out.  but it's also undeniable that it makes them easier to read for someone who doesn't know ahead of time how they're laid out.  it also makes them easier to read programmaticly, since I no longer have to parse everything.  and of course the fact that now I can rite tests for them is quite important.  so yeah, to it'll take some getting used, but I'm feeling good about this change.