linguistic tests
I wrote tests for the language stuff!  turns out it was buggy as all heck.  who would have gessed, with names like aooouwooh?  I had to revise the ± phoneme modification classes to be smarter, invert a few conditions in Harmony and Syllabicization, add more exceptions to some of the japanese words, add a new transcription flag that deals with nasalized vocoids, add some new rows to the alphabet.ts table, and modify a bunch of processes.  I also renamed SpecialProcess to PhraseProcess and Process to WordProcess because the old names were ridiculus.  wow, that was a slog.  but I'm almost done testing my resource files.  I just have to write something for resources/culture.ts now.