make cultures aware of biome and tidal locking
for a long time now I've bean meaning to set up the culture generatino so that if you live in a desert, you won't be expected to wear parkas or worship frogs or whatever.  similarly, ensuring that people on a tidally locked planet don't paint pictures of stars, since none of them would have ever seen a star.  it was sort of half-implemented for a long time, and in particular with this recent reformatting of the resource files it ended up being that any cultural feature that depended on location or planetary rotation was just never used.  no longer!  I had to revise the surface classes slitely so that a surface knows whether it has days, and reinstate the biome name list, but then it was pretty easy to just add location-based classes to the culture before it goes thru and selects all its feature values.