new surface tests
I moved surface.test.ts to voronoigraph.test.ts because it really is focused on the graph aspects of Surfaces and Vertexes and Tiles, and created a new surface.test.ts that goes into all of the geometric aspects of Surface's various subclasses.  issues of interest that I discoverd this way include the toroids being significantly smaller than they were supposed to be, the disc using its outer radius in a way that was harmless but inconsistent, docstrings and unit documentation being completely absent in places, and the computeEdgeVertexLocation function having an unused parameter. I also simplified the Surface constructor slitely since I realized it's extremely silly to specify whether the spheroid is rotating and also specify the rotation rate. better to just check whether the rotation rate is zero or not. the spheroid formula also had the wrong first-order coefficient.  I'm not sure how I got 1/2, but I found a textbook that derives it to be 4/5.  now the flattening with Earth's parameters matches the real Earth more closely.  it's still not a terribly *good* fit because I'm still assuming a uniformly dense planet, and Earth is not uniformly dense.  but as fun as it would be to read about those gigabar EOS tables I spend so much time helping PIs measure and compute a radial density profile given the total size, mass, and rotation rate, I think this is fine.