create unit tests for culture resource file
finally, I should stop getting surprise errors when a culture feature value key is absent in the translation files!  this also checks that the culture feature classes are all set up correctly, which they were not as it turns out.  the tech tree file was especially bad because I, uh, still haven't used it yet.  so it had an extra wrapping that would never have made sense with any code.  so I fixd that.