remove redundant enum lookup table
it's ever so slitely slower but much more robust to just iterate thru the enum keys and look for one that matches.  the previus version actually did have a bug, which is that it would have interpreted "breathy" as "stop" if it were ever in the first position in a phoneme specification.  it wasn't, but still better to avoid mistakes like that when I can.