move a function to avoid import issues
culture.ts is used by language.ts, which is used by civ.ts, which is used by world.ts, which is used by chart.ts, which is a problem because chart.ts uses internationalization.ts to format its culture strings, so I can't test chart.ts unless the internationalization strings are loaded.  this was an issue and quite silly to begin with, so I moved the function that formats the culture strings into factsheet.ts.