unify all map projections
I refactored the map projection code to eliminatethe need for different classes; they all now use the same equations – a generalized pseudoconic projection that can be parameterized to take the form of any cylindrical, pseudocylindrical, azimuthal, or conic projection I want.  I'm hoping this will streamline debugging even more so than the last commit, since now there's really only one projection about which to worry (as opposed to six!).  then agen, much like the commit before last, this has created some bugs that I'm not addressing just yet.  lots of roundoff garbage, basicly.  in my defense, these are problems that were always a little bit present, but now they're more noticeable because I've bilt more rigid code that can't compensate for small errors with as many context-dependent assumptions.  for example, itwas always the case that the antimeridian wouldn't *quite* line up on the conic projection when the surface is planar and the standard parallel is not 90°N.  but previusly if you were plotting a planar surface you were either using the azimuthal projection that *did* line up correctly, or you're not seeing the whole thing so it doesn't matter that it doesn't line up.  and that's still sort of the case now, but now the conic projection notices that it *should* line up at the antimeridian and thus assumes that it does, which becomes a problem because it in fact doesn't.  I swear to god I will get to a point where the map projections actually work for all landmasses on all surfaces.