move arability assignment to biome assignment
I moved the code where the tiles' arability and passability are calculated from the World constructor to the setBiomes() function.  it makes a bit more sense because othewise these are some of the only instance variables of Tile that get set in world.ts and not terrain.ts or surface.ts.  it also seems to fix a bug where the map regenerates with a seemingly different random seed when you change the inputs on the history tab for the first time, but I don't understand what it could have to do with that.  I'm not happy about the fact that I don't understand that.  it's really really hard to understand the inputs that go into these generation functions and how they arrive at the ansers at which they arrive, and I don't understand how much of that is just because it's a web-app and how much of that is my fault for somehow writing it poorly.