stop reading Edge.length before it's set
I found out why it affected the behavior to move the arability calculation to a different part of the code!  it's because that calculation reads the length of coastal Edges and that length isn't set when the Edge is constructed; by default it's only set when the Edge is greebled.  and whether it's greebled depends on whether a map has been made that uses that edge!  I set it to make sure it's always reading a real value and not just seeing undefined and casting it to 0.  thanks, JavaScript!