treat 'L' paths as rhumb-like, not great circles
the functions getGeoMidpoint, getMeridianCrossing, and getParallelCrossing collectively define the precise path that I assume "strait lines" in Paths to take thru geographic space for the purposes of plotting functions.  until now I've been trying to make them as close to great circles as possible, assuming that would look best.  it turns out, tho, that this is really hard.  in this update I change it to be the much simpler option of a strait line in latitude/longitude space, which may wrap around ф=π and/or λ=π if it would shorten the distance.  this has several advantages.  the main one is that its easier.  the other main one is that the previus version wasn't actually working, for a few reasons, such as that getGeoMidpoint, when operating on a sufficiently long line along the outer equator of a toroid, would return points on the inner equator, because it was trying to draw a strait line in 3D space and then snap that to the toroidal surface.  also, it was inconsistent, because the getParallelCrossing function was already (sort of) assuming the rhumb-like geometry, so some lines going thru multiple edges would have impossible sets of crossings