try to make cutToSize more robust
this one kind of ballooned out of my control.

my original goal here was just to redefine the conventions so that an empty path contains everything rather than containing noting.  that way when I want to describe the entire Surface I can pass `[]` to `cutToSize` or fill rather than `Chart.outline(surface)`.  it turns out, this is a hacky anser that works even worse in practice than the hacky anser I was using previusly.  because it means that if an island is off-screen, its path segments get removed when the screen is cut-to-size to the geographical region, and then when you go to cut-to-size to the rectangular region you see the nothing and think that means everything is included.  so an island off screen turns into the whole screen.

to fix that, I went ahead and implemented a more robust definition of what is inside a polygon and what is outside, and enhanced `cutToSize` to account for off-screen sections when determining if the region within the edges should be shaded.  meaning that for the first time it's actually not hacky and will always deal with off-screen entities correctly, rather than assuming them to be invisible as it did before or assuming them to be maximally visible as my original plan would have done.

in order to avoid making the `insideOut` function too complicated, I realized I could reuse `encompasses`, switching the region and the edges, as long as I expand it a little bit to take segments with parallels and meridians, rather than just linetos.  unfortunately the types didn't quite work out since the region is a list of path segments and the edges are a list of lists of the `MapEdge` interface.  so I had to scrap that interface.  it was really silly to have two ways of expressing the exact same information, anyway, so good riddance.  that also allowed me to use `getEdgeCrossing` in my `contains` function, saving me from having to write yet another function that accounts for two periodic dimensions.

well, except that `getEdgeCrossing` only returns *a* crossing, whereas I really need to find all crossings of a segment over a given set of edges.  so I then needed to refactor it to return multiple intersects.  only a small refactor, but still.

well, and except that `getEdgeCrossing` doesn't return the _polarity_ of the crossing, which is precisely what you need for a signed polygon bounding function like `contains`, so I had to edit it to include that information.  and doing that in conjunction with allowing parallels and meridians to feed into that function made it get really complicated, so I restructured it quite a bit to reuse more of that code (now getMeridianCrossing and getParallelCrossing are a single function).  and that included changing the function signature, in particular to remove the need to pass a whole starting segment (just the endpoint of the previus segment).  but the overall algorithm and behavior are the same.

well, except that `getEdgeCrossing` also ignores all intersections between the endpoint of a curve and an edge.  this is necessary with the old way I had ritten `cutToSize` because if I count any endpoint-edge intersections, I create an infinite loop wherein it repeatedly tries to splice the same crossing over and over agen, generating an infinite list of orphan moveto commands.  that doesn't work for `contains` because that will cause it to get its `crossingSum` rong when the `testEdge` goes thru a vertex.  so I had to edit `getEdgeCrossings` more substantially in order to get it to the level of quality I required from `contains`.  specificly, edit `getEdgeCrossings` to count endpoint-edge intersections iff the path segment is inside the edge (except the one point that's on the edge), but still ignore them when the path segment is entirely on or outside the edge.  this causes `contains` to always count each crossing once as it should, but requries a modification to `cutToSize` to kick it out of that infinite loop.  it also benefits `cutToSize` in the long run because it allows it to register the edge case that has long plaiged me where a path crosses the edge at a vertex that's exactly on it.

at some point my goal shifted to just making `contains` and `cutToSize` airtite.  but as it turns out, holy crap there are so many edge cases.  especially around vertices between edges, you need a ton of if cases to get everything just right.  so I'm intentionally ignoring the fact that it won't always get the periodicity right if a segment goes thru a corner that's on the antiequator or antimeridian, and the fact that it will sometimes see extra crossings if a segment is colinear with an edge and passes over one of its endpoints and the connecting edge at that endpoint happens to be facing the right direction.  if these things turn out to be issues, I think I should probably not fix them and instead go to an alternative implementation I thaut of just now where `getEdgeCrossings` is *only* used for `cutToSize` and is allowed to return as many extraneus crossings as it likes, and `contains` uses a different method based on purely vertical lines, but where the ray goes up from the point to infinity and then back down on an arbitrary segment of the polygon and thru to negative infinity, so that i don't have to worry about that elbow at all.  tho i suppose then the periodicity thing could still be an issue if there's a doubly-periodic double vertex.

but then!  finally!  after all that... it still works like fucking shit!  the rate of random errors doesn't seem to have changed at all!!!  I hate this code so much and I don't understand why it's being so unbearable but I feel like I have to commit something because it seems to be working about as well as it was before the refactor (:sob:) and I don't want this commit to get any bigger.